Gestion des erreurs + Fix connexion bd
-Ajout de la fermeture de connexion à la base de données après chaque fin de requête sur celle-ci

-Gestion des erreurs pour la suppression de bâtiments et de capteurs, cela affiche maintenant une erreur qui indique que le bâtiment ou capteur n'a pas été supprimé et redirige vers la page admin, en effet le script exécutait tout de même la requête sql de suppression lorsque l'on soumettait le formulaire avec de confirmation de suppression (radio button) sans faire de choix, ce qui en faisant un choix oui plus tard pose problème puisque l'id du capteur ou bâtiment a supprimer à été perdu par cause du formulaire vide soumis plutôt.

-Ajout de la fonctionnalité de filtres pour l'affichage des capteurs de la page admin, sachant que l'admin pourrait avoir énormément de capteurs à gérer

+++ public_html/delete_building.php
@@ -70,7 +70,7 @@
 
             <label for="non">No</label>
             <input type="radio" name="confirm" value="no" id="non">
-            
+        
             <input type="hidden" name="delete" value="<?php echo $id_building;?>">
 
             <input type="submit" value="VSubmit">
@@ -79,32 +79,44 @@
 
 
         <?php 
-
-            if (isset($_POST['confirm'])) {
+        
+            if ((!empty($_POST['confirm']) && ($_POST['confirm'] === "yes"))) {
                 
-                if ($_POST['confirm'] === "yes") {
-
-                    //Record the query that will delete this building from the database in the variable $query
+                if (is_numeric($_POST['delete']))
+                {
+                    //Record the query that will delete this sensor from the database in the variable $query
                     $query = "DELETE FROM building WHERE ID_building = " . $_POST['delete']; 
 
                     try {
                         mysqli_query($id_bd, $query);
                     } catch (Exception $e) {
-                        die("SQL REQUEST ERROR BUILDING NOT DELETED : <br>" . $e );
+                        mysqli_close($id_bd);
+                        die("SQL REQUEST ERROR BUILDING NOT REMOVED : <br>" . $e);
                     }
-
                     echo '
-                        <center> <h1> building successfully removed </h1> </center>
+                        <center> <h1> BUILDING SUCCESSFULLY REMOVED </h1> </center>
                         <script>
                             setTimeout(function() {
                                 window.location.href = "./admin.php";
                             }, 1500); 
                         </script>';
+                }else
+                {
+                    echo '
+                    <center> <h1> ERROR BUILDING NOT REMOVED </h1> </center>
+                    <script>
+                        setTimeout(function() {
+                            window.location.href = "./admin.php";
+                        }, 1500); 
+                    </script>';
                 }
-                elseif ($_POST['confirm'] === "no") {
-                    echo '<script> window.location.href = "./admin.php"; </script>';
-                }
-            }      
+
+            }elseif (!empty($_POST['confirm']) && ($_POST['confirm'] === "no"))
+            {
+                echo '<script> window.location.href = "./admin.php"; </script>';
+            }  
+                  
+            mysqli_close($id_bd);
             ?>
         </section>
     </body>
